+++ plugins/pycodetags_javascript/tests/test_import.py
@@ -1,0 +1,5 @@
+import pycodetags_javascript as the_module
+
+
+def test_import():
+    assert dir(the_module)