+++ plugins/pycodetags_to_sqlite/pycodetags_to_sqlite/main.py
@@ -1,0 +1,26 @@
+# code_tags_markdown_view.py
+import pluggy
+
+from pycodetags import DATA
+from pycodetags.config import CodeTagsConfig
+
+# Use the same marker name as defined in the host's hookspecs
+hookimpl = pluggy.HookimplMarker("pycodetags")
+
+
+@hookimpl
+def code_tags_print_report(
+    format_name: str,
+    found_data: list[DATA],
+    # pylint: disable=unused-argument
+    output_path: str,
+    # pylint: disable=unused-argument
+    config: CodeTagsConfig,
+) -> bool:
+    raise NotImplementedError()
+    # return False  # This plugin does not handle the requested format
+
+
+@hookimpl
+def code_tags_print_report_style_name() -> list[str]:
+    return ["markdown_simple"]