+++ pycodetags/folk_code_tags.py
@@ -28,7 +28,7 @@ try:
     from typing import Literal, TypedDict  # type: ignore[assignment,unused-ignore]
 except ImportError:
     from typing_extensions import Literal  # type: ignore[assignment,unused-ignore] # noqa
-    from typing_extensions import TypedDict # noqa
+    from typing_extensions import TypedDict  # noqa
 
 
 logger = logging.getLogger(__name__)