+++ tests/test_todo_skips.py
@@ -1,6 +1,0 @@
-import pytest
-
-
-@pytest.mark.skip(reason="no way of currently testing this")
-def test_todo_skips():
-    pass