+++ CHANGELOG.md
@@ -5,6 +5,37 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 2025-06-28
+
+## Changed
+- Huge refactoring into core library and plugins.
+
+## Removed
+- Issue tracker code. Now in pycodetags_issue_tracker.
+
+### Added
+- Several "application" plugins
+
+### Fixed 
+- Double code tags found when they match two schemas (folk and PEP350)
+
+
+## [0.2.0] - 2025-06-28
+
+## Added
+- Integration tests
+
+### Fixed 
+- Double code tags found when they match two schemas (folk and PEP350)
+
+## [0.1.2] - 2025-06-26
+
+## Fixed
+- Importing colored logging when it isn't available.
+- Defensive coding on arg parse
+- Defensive coding on "can't import"
+
+
 ## [0.1.1] - 2025-06-05
 
 ### Added
@@ -23,18 +54,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Plugins: Add plugin system
 - DATA: a code tag for supporting discussion, data, code review, etc.
 
-## [0.1.2] - 2025-06-26
 
-## Fixed
-- Importing colored logging when it isn't available.
-- Defensive coding on arg parse
-- Defensive coding on "can't import"
 
 ## [Unreleased]
 
-### Added
-- Views: Add dashboard website with filter, sort, detail views.
-- UnionListStr: Support values that are lists or strings
-
-### Fixed 
-- Double code tags found when they match two schemas (folk and PEP350)
+### Changes
+- Refactor issue tracker to a plugin to make core library strictly an abstract code tag (data tags)